Actually fix Demos && improve mem usage
This commit actually fixes demos without significantly increasing
the memory usage. To do this, demos are still stored in PROGMEM
but the correct methods are used to read the data.

Additionally, all of the log messages are also stored in PROGMEM,
which has a huge positive impact on the available memory but does
come with a bit of performance cost. As a result, the debug mode
should only be used for debugging and delays should not be less
than 200ms in this mode.